+++ meshtastenstein-core/src/ports/mod.rs
@@ -4,6 +4,7 @@ pub mod identity;
 pub mod reboot;
 pub mod sleep;
 pub mod storage;
+pub mod watchdog;
 
 pub use config_storage::ConfigStorage;
 pub use entropy::EntropySource;
@@ -11,6 +12,7 @@ pub use identity::Identity;
 pub use reboot::Reboot;
 pub use sleep::Sleep;
 pub use storage::{Storage, StorageError};
+pub use watchdog::Watchdog;
 
 pub trait MeshStorage: ConfigStorage + storage::Storage {}
 impl<T: ConfigStorage + storage::Storage> MeshStorage for T {}